badcontrols: Difference-in-Differences with Bad Controls

Implements methods for difference-in-differences when time-varying covariates are affected by treatment (bad controls). Provides imputation-based and doubly robust machine learning estimators that correctly handle post-treatment covariates. Builds on the panel treatment effects infrastructure in 'ptetools'. Based on Caetano, Callaway, Payne, and Sant'Anna (2024) "Difference-in-Differences with Bad Controls".

See also

Author

Maintainer: Hugo Sant'Anna hsantanna@uab.edu

Authors:

Hugo Sant'Anna hsantanna@uab.edu

Carolina Caetano ccaetano@uga.edu

Brantly Callaway brantly.callaway@uga.edu

Stroud Payne stroud.payne@vanderbilt.edu